Save game result & refactor dataset gen
