+++ src/nnue.h
@@ -4,7 +4,7 @@
 #include "board.h"
 
 #define INPUTS_COUNT 768
-#define INNER_LAYER_COUNT 8
+#define INNER_LAYER_COUNT 256
 #define MAX_FEN_LENGTH 1000
 
 
@@ -37,9 +37,9 @@ void genDataset(Board* board, int from, int to, char* filename);
 void checkAndSavePositionToDataset(Board* board, TimeManager tm, FILE* file);
 int saveQuiet(Board* board, TimeManager tm, FILE* file, char* fen);
 
-void createDataset(Board* board, int gamesCount, int seed, char* fileName, char* logFile);
-void runGame(Board* board, FILE* file);
-void makeRandomMove(Board* board);
-int getMovesCount(Board* board);
+//void createDataset(Board* board, int gamesCount, int seed, char* fileName, char* logFile);
+//void runGame(Board* board, FILE* file);
+//void makeRandomMove(Board* board);
+//int getMovesCount(Board* board);
 
 #endif
