+++ src/uci.c
@@ -1,4 +1,5 @@
 #include "uci.h"
+#include "dataset.h"
 
 char startpos[] = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
 Option option;
